+++ src/oci/core/models/dedicated_vm_host.py
@@ -107,6 +107,14 @@ class DedicatedVmHost(object):
             The value to assign to the remaining_memory_in_gbs property of this DedicatedVmHost.
         :type remaining_memory_in_gbs: float
 
+        :param capacity_bins:
+            The value to assign to the capacity_bins property of this DedicatedVmHost.
+        :type capacity_bins: list[oci.core.models.CapacityBin]
+
+        :param compute_bare_metal_host_id:
+            The value to assign to the compute_bare_metal_host_id property of this DedicatedVmHost.
+        :type compute_bare_metal_host_id: str
+
         """
         self.swagger_types = {
             'availability_domain': 'str',
@@ -123,7 +131,9 @@ class DedicatedVmHost(object):
             'total_ocpus': 'float',
             'remaining_ocpus': 'float',
             'total_memory_in_gbs': 'float',
-            'remaining_memory_in_gbs': 'float'
+            'remaining_memory_in_gbs': 'float',
+            'capacity_bins': 'list[CapacityBin]',
+            'compute_bare_metal_host_id': 'str'
         }
         self.attribute_map = {
             'availability_domain': 'availabilityDomain',
@@ -140,7 +150,9 @@ class DedicatedVmHost(object):
             'total_ocpus': 'totalOcpus',
             'remaining_ocpus': 'remainingOcpus',
             'total_memory_in_gbs': 'totalMemoryInGBs',
-            'remaining_memory_in_gbs': 'remainingMemoryInGBs'
+            'remaining_memory_in_gbs': 'remainingMemoryInGBs',
+            'capacity_bins': 'capacityBins',
+            'compute_bare_metal_host_id': 'computeBareMetalHostId'
         }
         self._availability_domain = None
         self._compartment_id = None
@@ -157,6 +169,8 @@ class DedicatedVmHost(object):
         self._remaining_ocpus = None
         self._total_memory_in_gbs = None
         self._remaining_memory_in_gbs = None
+        self._capacity_bins = None
+        self._compute_bare_metal_host_id = None
 
     @property
     def availability_domain(self):
@@ -581,6 +595,54 @@ class DedicatedVmHost(object):
         :type: float
         """
         self._remaining_memory_in_gbs = remaining_memory_in_gbs
+
+    @property
+    def capacity_bins(self):
+        """
+        Gets the capacity_bins of this DedicatedVmHost.
+        A list of total and remaining CPU & memory per capacity bucket.
+
+
+        :return: The capacity_bins of this DedicatedVmHost.
+        :rtype: list[oci.core.models.CapacityBin]
+        """
+        return self._capacity_bins
+
+    @capacity_bins.setter
+    def capacity_bins(self, capacity_bins):
+        """
+        Sets the capacity_bins of this DedicatedVmHost.
+        A list of total and remaining CPU & memory per capacity bucket.
+
+
+        :param capacity_bins: The capacity_bins of this DedicatedVmHost.
+        :type: list[oci.core.models.CapacityBin]
+        """
+        self._capacity_bins = capacity_bins
+
+    @property
+    def compute_bare_metal_host_id(self):
+        """
+        Gets the compute_bare_metal_host_id of this DedicatedVmHost.
+        The compute bare metal host OCID of the dedicated virtual machine host.
+
+
+        :return: The compute_bare_metal_host_id of this DedicatedVmHost.
+        :rtype: str
+        """
+        return self._compute_bare_metal_host_id
+
+    @compute_bare_metal_host_id.setter
+    def compute_bare_metal_host_id(self, compute_bare_metal_host_id):
+        """
+        Sets the compute_bare_metal_host_id of this DedicatedVmHost.
+        The compute bare metal host OCID of the dedicated virtual machine host.
+
+
+        :param compute_bare_metal_host_id: The compute_bare_metal_host_id of this DedicatedVmHost.
+        :type: str
+        """
+        self._compute_bare_metal_host_id = compute_bare_metal_host_id
 
     def __repr__(self):
         return formatted_flat_dict(self)